feat: Add props parameter to FsxLifecycleStatusMonitor constructor
This commit adds a new optional `props` parameter to the `FsxLifecycleStatusMonitor` constructor in order to provide additional properties for the FSx Lifecycle Status Monitor. The `props` parameter is of type `FsxLifecycleStatusMonitorProps`. This change allows for more flexibility and customization when creating an instance of the monitor.

The commit also includes updates to the API documentation, specifically adding a new struct called `FsxLifecycleStatusMonitorProps`, which contains properties for the FSx Lifecycle Status Monitor. Additionally, a new constant called `DEFAULT_SCHEDULE` has been added to the monitor class.

In the integration tests, this commit modifies the instantiation of `FsxLifecycleStatusMonitor` by passing an empty object as the value for the new `props` parameter. It also updates some test assertions and snapshot files accordingly.